Performed basic housekeeping on AliasChangeLog, CalcRankScore, Character, CharacterChangeLog. I will cover the rest at a later date. No new implementations.
Note:
Changelogs do not need to have the same constraints as the tables they are logging. Those tables are doing that for them already.

That said, you still need to make the proper constraints for both PKs and FKs. Also, don't forget to create the nonclustered indexes. Shortcuts now are only going to hurt us later.